requirements.txt
click==8.1.7
openai==1.10.0
python-dotenv==1.0.1
streamlit==1.30.0
tqdm==4.66.1
